Several changes:
Made "to personal" button for search result pivot view only visible when this view is selected.
Tap personal pin or search result in manage page to view details.
Fixed the bug that after going back to map view by tapping the "go to" button of a personal pin or search result in the manage page, tapping outside the expanded pin did not collapse the pin.